+++ merlin/router.py
@@ -211,8 +211,9 @@ def create_config(task_server: str, config_dir: str, broker: str) -> None:
     """
     Create a config for the given task server.
 
-    :param `task_server`: The task server from which to stop workers.
-    :param `config_dir`: Optional directory to install the config.
+    :param [str] `task_server`: The task server from which to stop workers.
+    :param [str] `config_dir`: Optional directory to install the config.
+    :param [str] `broker`: string indicated the broker, used to check for redis.
     """
     LOG.info("Creating config ...")